requirements.txt
torch==2.3.0+cu121          --extra-index-url https://download.pytorch.org/whl/cu121
xformers==0.0.26.post1      --extra-index-url https://download.pytorch.org/whl/cu121
diffusers==0.27.2
transformers==4.41.2
accelerate==0.30.1
safetensors==0.4.3
numpy==1.26.4
scipy==1.12.0
Pillow==10.3.0
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1